pkg/helm: move ownerref-injecting client to client package
